Default OpenDoc UniRec recognition to FP16

+++ 12_unirec_0_1b_inference/run_opendoc_custom_unirec.py
@@ -39,7 +39,7 @@ def parse_args() -> argparse.Namespace:
     parser.add_argument(
         "--dtype",
         choices=("bfloat16", "float16", "float32"),
-        default="bfloat16",
+        default="float16",
     )
     parser.add_argument("--max-length", type=int, default=2048)
     parser.add_argument(
